requirements.txt
cassio 
datasets 
langchain 
openai 
tiktoken
streamlit
PyPDF2
Pillow
python-dotenv
langchain-community